fix warnings on Mac in src/gpu and include/gpu
Fix these class of warnings:
- unused functions
- unused locals
- sign mismatch
- missing function prototypes
- missing newline at end of file
- 64 to 32 bit truncation

The changes prefer to link in dead code in the debug build
with 'if (false)' than to comment it out, but trivial cases
are commented out or sometimes deleted if it appears to be
a copy/paste error.
Review URL: https://codereview.appspot.com/6296044

git-svn-id: http://skia.googlecode.com/svn/trunk@4180 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/gpu/GrTexture.cpp
@@ -124,8 +124,8 @@ void gen_texture_key_values(const GrGpu* gpu,
     // assert that texture creation will fail anyway if this assumption
     // would cause key collisions.
     GrAssert(gpu->getCaps().fMaxTextureSize <= SK_MaxU16);
-    v[0] = clientKey & 0xffffffffUL;
-    v[1] = (clientKey >> 32) & 0xffffffffUL;
+    v[0] = (uint32_t) (clientKey & 0xffffffffUL);
+    v[1] = (uint32_t) ((clientKey >> 32) & 0xffffffffUL);
     v[2] = desc.fWidth | (desc.fHeight << 16);
 
     v[3] = (desc.fSampleCnt << 24);
